+++ test/Media-test.js
@@ -19,31 +19,14 @@ describe("Media", () => {
   });
 
   it("renders with default style", () => {
-    assert.deepEqual(result.props.style, {
-      marginTop: "1.5em",
-      marginBottom: "1.5em"
-    });
+    assert.deepEqual(result.props.style, {});
   });
 
   it("when given a style prop, it merges styles", () => {
     shallowRenderer.render(<Media style={{ backgroundColor: "blue" }}><div /></Media>);
     result = shallowRenderer.getRenderOutput();
 
-    assert.deepEqual(result.props.style, {
-      backgroundColor: "blue",
-      marginTop: "1.5em",
-      marginBottom: "1.5em"
-    });
-  });
-
-  it("when given a style prop, with a rule that overrides a default, it replaces the default", () => {
-    shallowRenderer.render(<Media style={{ marginTop: "3em" }}><div /></Media>);
-    result = shallowRenderer.getRenderOutput();
-
-    assert.deepEqual(result.props.style, {
-      marginTop: "3em",
-      marginBottom: "1.5em"
-    });
+    assert.deepEqual(result.props.style, { backgroundColor: "blue" });
   });
 
   it("has a styled `:before`-like element", () => {